CRM/requirements.txt
albumentations==0.4.3
opencv-python==4.9.0.80
pudb==2019.2
imageio==2.34.1
imageio-ffmpeg==0.4.2
pytorch-lightning==1.4.2
torchmetrics==0.6
omegaconf==2.3.0
test-tube==0.7.5
streamlit==1.34.0
einops==0.7.0
transformers==4.27.1
webdataset==0.2.5
open-clip-torch==2.7.0
gradio==4.29.0
kornia==0.6
invisible-watermark==0.2.0
streamlit-drawable-canvas==0.9.3
ninja==1.11.1.1
nvdiffrast==0.3.1
diffusers==0.24.0
huggingface-hub==0.22.2
Pillow==10.1.0
opencv-contrib-python-headless==4.9.0.80
opencv-python-headless==4.9.0.80
rembg==2.0.56
pygltflib==1.16.2
kiui==0.2.8
trimesh==4.3.1
xatlas==0.0.9
pymeshlab==2023.12.post1